+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Arbitrary.hs
@@ -32,7 +32,8 @@ import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Binary (EncCBOR (..))
 import Cardano.Ledger.Shelley (ShelleyEra)
 import Cardano.Ledger.Shelley.API (
-  ApplyTxError (ApplyTxError),
+  ApplyTx (..),
+  ApplyTxError (..),
   MultiSig,
   ShelleyDelegCert,
   ShelleyTx (ShelleyTx),
@@ -712,14 +713,7 @@ instance
 
 deriving newtype instance Arbitrary (Tx TopTx ShelleyEra)
 
-instance
-  ( Era era
-  , Arbitrary (PredicateFailure (EraRule "LEDGER" era))
-  ) =>
-  Arbitrary (ApplyTxError era)
-  where
-  arbitrary = ApplyTxError <$> arbitrary
-  shrink (ApplyTxError xs) = [ApplyTxError xs' | xs' <- shrink xs]
+deriving newtype instance Arbitrary (ApplyTxError ShelleyEra)
 
 instance
   ( Era era